+++ MindTheWorld.Database/Tables/EpidemicDeath.sql
@@ -2,7 +2,7 @@
 (
 	[CountryId] INT NOT NULL , 
     [Year] INT NOT NULL, 
-    [Value] INT NULL, 
+    [Value] FLOAT NULL, 
 
     CONSTRAINT PK_ED_ID PRIMARY KEY ([CountryId], [Year]),
     CONSTRAINT FK_ED_CountryId FOREIGN KEY (CountryId) REFERENCES Country(Id)
