+++ MindTheWorld.Database/Tables/GrossDomesticProduct.sql
@@ -2,7 +2,7 @@
 (
 	[CountryId] INT NOT NULL, 
     [Year] INT NOT NULL, 
-    [Value] DECIMAL(10, 6) NULL, 
+    [Value] FLOAT NULL, 
 
     CONSTRAINT PK_GDP_ID PRIMARY KEY ([CountryId], [Year]),
     CONSTRAINT FK_GDP_CountryId FOREIGN KEY (CountryId) REFERENCES Country(Id)
